Add header.css some style
