+++ src/App.js
@@ -1,4 +1,3 @@
-import './App.css';
 import Header from './components/Header';
 
 function App() {
